move tree2net from feature to complex
