+++ knowde/complex/tree2net/__init__.py
@@ -2,10 +2,11 @@
 from __future__ import annotations
 
 from knowde.complex.__core__.sysnet import SysNet
-from knowde.feature.parser.tree2net.interpreter import SysNetInterpreter
-from knowde.feature.parser.tree2net.transformer import TSysArg
 from knowde.primitive.parser import parse2tree
 from knowde.primitive.parser.testing import treeprint
+
+from .interpreter import SysNetInterpreter
+from .transformer import TSysArg
 
 
 def parse2net(txt: str, do_print: bool = False) -> SysNet:  # noqa: FBT001 FBT002
