+++ knowde/complex/tree2net/ranking/__init__.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel
 
 from knowde.complex.__core__.sysnet import SysNet
 from knowde.complex.__core__.sysnet.sysnode import SysArg
-from knowde.feature.parser.tree2net.ranking.node_index import (
+from knowde.complex.tree2net.ranking.node_index import (
     Conclusion,
     Description,
     Premise,
