+++ knowde/complex/tree2net/test_ranking.py
@@ -1,7 +1,7 @@
 """テストケース作るの楽."""
 
 
-from knowde.feature.parser.tree2net import parse2net
+from knowde.complex.tree2net import parse2net
 
 
 def test_ranking() -> None:
